requirements.txt
absl-py==0.11.0
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto3==1.21.26
botocore==1.24.26
cachetools==4.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==8.0.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.0
jmespath==1.0.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.14.0
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pymongo==4.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.5.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.12.1
zipp==3.7.0
